+++ CHANGELOG.md
@@ -5,6 +5,7 @@ In order to read more about upgrading and BC breaks have a look at the [UPGRADE 
 
 ## 2.2.0
 
++ [#166](https://github.com/luyadev/luya-module-cms/issues/166) New LanguageSwitcher method `setUrlRuleParam()` to make url rules translatable.
 + [#224](https://github.com/luyadev/luya-module-cms/pull/224) Add `getBlock()` method in PhpBlockView to access block context.
 
 ## 2.1.1 (17. September 2019)